Fix/38233 webauth preload fido (#38664)
* Use cached FIDO metadata

Places a copy of fido.jwt in the plugin's directory

* Use cached FIDO metadata

Use the cached FIDO metadata in the plugin

* Use cached FIDO metadata

Remove unused local file fallback

* Use cached FIDO metadata

Simplify

* Use cached FIDO metadata

Update Joomla build script

* Use cached FIDO metadata

Let's not break the build if the FIDO server is down

* Use cached FIDO metadata

Do not include the fido.jwt in the repo

* Use cached FIDO metadata

Make sure not to re-fetch the fido.jwt file if it already exists.

* Use cached FIDO metadata

Move the check for the fido.jwt file right after composer install, do not try to fetch the file twice

* Use cached FIDO metadata

Update lang string

* Use cached FIDO metadata

Set the attestation support to disabled by default

* Update deleted files to PR #38664

+++ build/build.php
@@ -270,6 +270,14 @@ if ($composerReturnCode !== 0) {
     exit(1);
 }
 
+// Try to update the fido.jwt file
+if (!file_exists(rtrim($fullpath, '\\/') . '/plugins/system/webauthn/fido.jwt'))
+{
+    echo "The file plugins/system/webauthn/fido.jwt was not created. Build failed.\n";
+
+    exit (1);
+}
+
 system('npm install --unsafe-perm', $npmReturnCode);
 
 if ($npmReturnCode !== 0) {
